requirements.txt
fastapi
uvicorn
docxtpl
python-docx
python-dotenv
